refactor: refactor game code
