+++ react-game/src/components/chooseTheGame/gameItem/item.module.scss
@@ -7,6 +7,9 @@
 .beforeClick{
   border: 2px solid black;
   background-color: grey;
+  font-size: 40px;
+  font-weight: 600;
+  color: black;
 }
 
 .friendly {
